feat: implement cross-platform file system listing with compact, human-readable formatting and summary stats
