+++ src/cmds/system/ls_win.rs
@@ -1,6 +1,5 @@
-use crate::cmds::system::ls_format::LsRecordType;
 use super::constants::NOISE_DIRS;
-use super::ls::{self, LsRecord};
+use super::ls::{self, LsRecord, LsRecordType};
 
 use anyhow::Result;
 use colored::Colorize; 
@@ -9,6 +8,14 @@ use std::io::IsTerminal;
 use std::path::Path;
 use std::time::UNIX_EPOCH;
 
+pub fn estimate_raw_dir_output(records: &[LsRecord]) -> String {
+    let mut chars = 8; // "total 0\n"
+    for r in records {
+        // Heuristic: ~50 chars of fixed `ls -la` metadata overhead + filename length
+        chars += 45 + r.name.len();
+    }
+    " ".repeat(chars)
+}
 
 /// Fetches file information from the filesystem using native Rust std::fs.
 /// refactor required
@@ -131,7 +138,7 @@ fn warn_unsupported_flags(flags: &[String]) {
     }
 }
 
-pub fn run_native(paths: Vec<String>, show_all: bool, flags: Vec<String>) -> Result<(i32, String)> {
+pub fn run_native(paths: Vec<String>, show_all: bool, flags: Vec<String>) -> Result<(i32, String, String)> {
     warn_unsupported_flags(&flags);
 
     let active_flags: HashSet<char> = flags
@@ -160,7 +167,9 @@ pub fn run_native(paths: Vec<String>, show_all: bool, flags: Vec<String>) -> Res
         records.reverse();
     }
 
-    let (entries, summary) = ls::synthesize_output(records);
+    let raw_estimate = estimate_raw_dir_output(&records);
+
+    let (entries, summary) = super::ls_format::synthesize_output(records);
     let is_tty = std::io::stdout().is_terminal();
     let output = if is_tty {
         format!("{}{}", entries, summary)
@@ -168,5 +177,134 @@ pub fn run_native(paths: Vec<String>, show_all: bool, flags: Vec<String>) -> Res
         entries
     };
 
-    Ok((0, output))
+    Ok((0, output, raw_estimate))
 }
+
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use tempfile::tempdir;
+    use std::fs::{self, File};
+    use crate::cmds::system::ls_format::synthesize_output;
+
+    #[test]
+    fn test_fetch_entries_basic() {
+        let dir = tempdir().unwrap();
+        let dir_path = dir.path();
+
+        fs::create_dir(dir_path.join("src")).unwrap();
+        File::create(dir_path.join("Cargo.toml")).unwrap();
+        File::create(dir_path.join("README.md")).unwrap();
+
+        let records = fetch_entries(&[dir_path.to_string_lossy().into_owned()], false).unwrap();
+        let (entries, _summary) = synthesize_output(records);
+
+        assert!(entries.contains("src/"));
+        assert!(entries.contains("Cargo.toml"));
+        assert!(entries.contains("README.md"));
+        assert!(!entries.contains("total"));
+    }
+
+    #[test]
+    fn test_fetch_entries_filters_noise() {
+        let dir = tempdir().unwrap();
+        let dir_path = dir.path();
+
+        fs::create_dir(dir_path.join("node_modules")).unwrap();
+        fs::create_dir(dir_path.join(".git")).unwrap();
+        fs::create_dir(dir_path.join("target")).unwrap();
+        fs::create_dir(dir_path.join("src")).unwrap();
+        File::create(dir_path.join("main.rs")).unwrap();
+
+        let records = fetch_entries(&[dir_path.to_string_lossy().into_owned()], false).unwrap();
+        let (entries, _summary) = synthesize_output(records);
+
+        assert!(!entries.contains("node_modules"));
+        assert!(!entries.contains(".git"));
+        assert!(!entries.contains("target"));
+        assert!(entries.contains("src/"));
+        assert!(entries.contains("main.rs"));
+    }
+
+    #[test]
+    fn test_fetch_entries_show_all() {
+        let dir = tempdir().unwrap();
+        let dir_path = dir.path();
+
+        fs::create_dir(dir_path.join(".git")).unwrap();
+        fs::create_dir(dir_path.join("src")).unwrap();
+
+        let records = fetch_entries(&[dir_path.to_string_lossy().into_owned()], true).unwrap();
+        let (entries, _summary) = synthesize_output(records);
+
+        assert!(entries.contains(".git/"));
+        assert!(entries.contains("src/"));
+    }
+
+    #[test]
+    fn test_fetch_entries_empty() {
+        let dir = tempdir().unwrap();
+        let dir_path = dir.path();
+
+        let records = fetch_entries(&[dir_path.to_string_lossy().into_owned()], false).unwrap();
+        let (entries, summary) = synthesize_output(records);
+
+        assert_eq!(entries, "(empty)\n");
+        assert!(summary.is_empty());
+    }
+
+    #[test]
+    fn test_fetch_entries_symlinks() {
+        let dir = tempdir().unwrap();
+        let dir_path = dir.path();
+
+        let target_path = dir_path.join("target.txt");
+        File::create(&target_path).unwrap();
+
+        let link_path = dir_path.join("link.txt");
+        
+        // Handling symlinks specifically on windows 
+        #[cfg(windows)]
+        let symlink_result = std::os::windows::fs::symlink_file(&target_path, &link_path);
+        
+        #[cfg(not(windows))]
+        let symlink_result = std::os::unix::fs::symlink(&target_path, &link_path);
+
+        if let Err(e) = symlink_result {
+            // Ignore error if symlink creation failed due to lack of administrative privileges on Windows
+            eprintln!("Failed to create symlink: {:?}", e);
+            return;
+        }
+
+        let records = fetch_entries(&[dir_path.to_string_lossy().into_owned()], false).unwrap();
+        let (entries, _summary) = synthesize_output(records);
+
+        // NOTE: This assertion will FAIL with your current code, as fetch_entries 
+        // does not append " -> target.txt" to the file name. 
+        assert!(
+            entries.contains("link.txt -> target.txt") || entries.contains(&format!("link.txt -> {}", target_path.to_string_lossy())),
+            "Symlink output does not include target, got entries:\n{}", 
+            entries
+        );
+    }
+
+    #[test]
+    fn test_fetch_entries_summary() {
+        let dir = tempdir().unwrap();
+        let dir_path = dir.path();
+
+        fs::create_dir(dir_path.join("src")).unwrap();
+        File::create(dir_path.join("main.rs")).unwrap();
+        File::create(dir_path.join("lib.rs")).unwrap();
+        File::create(dir_path.join("Cargo.toml")).unwrap();
+
+        let records = fetch_entries(&[dir_path.to_string_lossy().into_owned()], false).unwrap();
+        let (_entries, summary) = synthesize_output(records);
+
+        assert!(summary.contains("Summary: 3 files, 1 dirs"));
+        assert!(summary.contains(".rs"));
+        assert!(summary.contains(".toml"));
+    }
+}
+
